merging

+++ backend/src/main/java/com/jangbo/db/repository/StoreRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoreRepository extends JpaRepository<Store,Long> {
+
+public interface StoreRepository extends JpaRepository<Store,Integer> {
 
 }
